requirements.txt
pydantic>=2.7
pyyaml>=6.0
tqdm>=4.66
pandas>=2.2
requests>=2.32
